backend/manage.py
#!/usr/bin/env python
import os
import sys
from pathlib import Path

if __name__ == "__main__":
    # Load .env file so environment variables are available to Django settings
    try:
        from dotenv import load_dotenv

        load_dotenv(Path(__file__).resolve().parent / ".env")
    except ImportError:
        pass

    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "digitn.settings")
    try:
        from django.core.management import execute_from_command_line
    except ImportError as exc:
        raise ImportError("Couldn't import Django.") from exc
    execute_from_command_line(sys.argv)
